+++ client/smart-contract/src/main/java/hera/contract/ClientPrepareable.java
@@ -1,0 +1,9 @@
+package hera.contract;
+
+import hera.client.AergoClient;
+
+interface ClientPrepareable {
+
+  void prepareClient(AergoClient aergoClient);
+
+}