+++ client/smart-contract/src/main/java/hera/contract/SignerPreparable.java
@@ -1,0 +1,13 @@
+/*
+ * @copyright defined in LICENSE.txt
+ */
+
+package hera.contract;
+
+import hera.key.Signer;
+
+interface SignerPreparable {
+
+  void prepareSigner(Signer signer);
+
+}