+++ client/smart-contract/src/test/java/hera/contract/ContractApiFactoryTest.java
@@ -8,25 +8,22 @@ import static org.junit.Assert.assertNotNull;
 
 import hera.AbstractTestCase;
 import hera.api.model.ContractAddress;
+import hera.key.AergoKeyGenerator;
 import org.junit.Test;
 
 public class ContractApiFactoryTest extends AbstractTestCase {
 
-  protected final ContractAddress contractAddress = ContractAddress
-      .of("AmgSvgaUXFb4PFLn5MU5wiYNihcQrwNhNMzaG5h68q99gBNTTkWK");
-
   @Test
   public void testCreate() {
+    final ContractAddress anyContractAddress = new AergoKeyGenerator().create().getAddress()
+        .adapt(ContractAddress.class);
     final ContractApi<TestContract> contractApi = new ContractApiFactory()
-        .create(contractAddress, TestContract.class, getClass().getClassLoader());
+        .create(anyContractAddress, TestContract.class);
     assertNotNull(contractApi);
   }
 
-  public static interface TestContract {
+  private interface TestContract {
 
-    void execute();
-
-    String query();
   }
 
 }