+++ client/smart-contract/src/test/java/hera/contract/ContractApiImplTest.java
@@ -5,40 +5,33 @@
 package hera.contract;
 
 import static org.junit.Assert.assertNotNull;
+import static org.mockito.Mockito.mock;
 
 import hera.AbstractTestCase;
-import hera.api.model.Fee;
-import hera.keystore.InMemoryKeyStore;
-import hera.wallet.WalletApi;
-import hera.wallet.WalletApiFactory;
+import hera.client.AergoClient;
+import hera.key.AergoKeyGenerator;
+import hera.key.Signer;
 import org.junit.Test;
 
 public class ContractApiImplTest extends AbstractTestCase {
+
+  @Test
+  public void testStep() {
+    final ContractApi<ContractTest> contractApi =
+        new ContractApiImpl<ContractTest>(new ContractTestImpl(), mock(ContractInvocationHandler.class));
+    final Signer signer = new AergoKeyGenerator().create();
+    ContractTest proxy1 = contractApi.with(mock(AergoClient.class)).execution(signer);
+    ContractTest proxy2 = contractApi.with(mock(AergoClient.class)).query();
+    assertNotNull(proxy1);
+    assertNotNull(proxy2);
+  }
 
   private interface ContractTest {
 
   }
 
-  private class ContractTestImpl implements ContractTest, ContractInvocationPreparable {
+  private class ContractTestImpl implements ContractTest {
 
-    @Override
-    public void setWalletApi(WalletApi walletApi) {}
-
-    @Override
-    public void setFee(Fee fee) {}
-
-  }
-
-  @Test
-  public void testStep() {
-    final ContractApi<ContractTest> contractApi =
-        new ContractApiImpl<ContractTest>(new ContractTestImpl());
-    final WalletApi walletApi = new WalletApiFactory().create(new InMemoryKeyStore());
-    final Fee fee = Fee.ZERO;
-    final ContractTest withFee = contractApi.walletApi(walletApi).fee(fee);
-    final ContractTest withoutFee = contractApi.walletApi(walletApi).noFee();
-    assertNotNull(withFee);
-    assertNotNull(withoutFee);
   }
 
 }