+++ client/smart-contract/src/test/java/hera/contract/ContractInvocationHandlerTest.java
@@ -6,142 +6,258 @@ package hera.contract;
 
 import static java.util.Arrays.asList;
 import static java.util.UUID.randomUUID;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import hera.AbstractTestCase;
 import hera.api.ContractOperation;
-import hera.api.model.Authentication;
 import hera.api.model.BytesValue;
 import hera.api.model.ContractAddress;
 import hera.api.model.ContractFunction;
 import hera.api.model.ContractInterface;
 import hera.api.model.ContractInvocation;
 import hera.api.model.ContractResult;
-import hera.api.model.ContractTxHash;
 import hera.api.model.Fee;
 import hera.api.model.StateVariable;
+import hera.api.model.TxHash;
 import hera.client.AergoClient;
-import hera.contract.ContractInvocationHandler;
-import hera.contract.ContractInvocationPreparable;
-import hera.key.AergoKey;
+import hera.client.AergoClientBuilder;
+import hera.client.TxRequestFunction;
+import hera.client.TxRequester;
 import hera.key.AergoKeyGenerator;
 import hera.key.Signer;
-import hera.keystore.InMemoryKeyStore;
-import hera.model.KeyAlias;
-import hera.wallet.WalletApi;
-import hera.wallet.WalletApiFactory;
-import java.lang.reflect.InvocationHandler;
+import java.io.IOException;
 import java.lang.reflect.Proxy;
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import org.junit.Test;
 
 public class ContractInvocationHandlerTest extends AbstractTestCase {
 
-  private interface ContractTest extends ContractInvocationPreparable {
-
-    void execute(int arg);
-
-    int query();
-
-  }
-
-  private final ContractAddress contractAddress = ContractAddress
-      .of("AmgSvgaUXFb4PFLn5MU5wiYNihcQrwNhNMzaG5h68q99gBNTTkWK");
+  private final ContractAddress contractAddress = new AergoKeyGenerator().create().getAddress()
+      .adapt(ContractAddress.class);
 
   private final ClassLoader classLoader = getClass().getClassLoader();
 
   @Test
-  public void testPrepare() {
-    final InvocationHandler invocationHandler = new ContractInvocationHandler(contractAddress);
-    ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
-        new Class<?>[]{ContractTest.class}, invocationHandler);
-
-    final WalletApi walletApi = supplyWalletApi();
-    final Fee fee = Fee.ZERO;
-    proxy.setWalletApi(walletApi);
-    proxy.setFee(fee);
+  public void testPrepareClient() {
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mock(
+        TxRequester.class));
+    final AergoClient aergoClient = new AergoClientBuilder().build();
+    invocationHandler.prepareClient(aergoClient);
   }
 
   @Test
-  public void testWithReturn() {
-    final InvocationHandler invocationHandler = new ContractInvocationHandler(contractAddress);
-    ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
-        new Class<?>[]{ContractTest.class}, invocationHandler);
-
-    final WalletApi walletApi = supplyWalletApi();
-    proxy.setWalletApi(walletApi);
-    proxy.setFee(Fee.ZERO);
-    proxy.execute(0);
+  public void testPrepareSigner() {
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mock(
+        TxRequester.class));
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareSigner(signer);
   }
 
   @Test
-  public void testWithoutReturn() {
-    final InvocationHandler invocationHandler = new ContractInvocationHandler(contractAddress);
-    ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+  public void testExecuteWithoutReturn() throws Exception {
+    // given
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final TxHash expected = TxHash.of(BytesValue.of(randomUUID().toString().getBytes()));
+    when(mockTxRequester
+        .request(any(AergoClient.class), any(Signer.class), any(TxRequestFunction.class)))
+        .thenReturn(expected);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
+
+    // then
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
         new Class<?>[]{ContractTest.class}, invocationHandler);
-
-    final WalletApi walletApi = supplyWalletApi();
-    proxy.setWalletApi(walletApi);
-    proxy.query();
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+    proxy.executeVoid();
   }
 
-  protected WalletApi supplyWalletApi() {
-    final InMemoryKeyStore keyStore = new InMemoryKeyStore();
-    final KeyAlias alias = KeyAlias.of(randomUUID().toString().replaceAll("-", ""));
-    final String password = randomUUID().toString();
-    final Authentication authentication = Authentication.of(alias, password);
-    final AergoKey key = new AergoKeyGenerator().create();
-    keyStore.save(authentication, key);
+  @Test
+  public void testExecuteWithTxHashReturn() throws Exception {
+    // given
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final TxHash expected = TxHash.of(BytesValue.of(randomUUID().toString().getBytes()));
+    when(mockTxRequester
+        .request(any(AergoClient.class), any(Signer.class), any(TxRequestFunction.class)))
+        .thenReturn(expected);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
 
-    final WalletApi walletApi = new WalletApiFactory().create(keyStore);
-    walletApi.bind(supplyAergoClient());
-    walletApi.unlock(authentication);
-
-    return walletApi;
+    // then
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+        new Class<?>[]{ContractTest.class}, invocationHandler);
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+    final TxHash actual = proxy.executeTxHash(randomUUID().toString(), randomUUID().toString());
+    assertEquals(expected, actual);
   }
 
-  @SuppressWarnings("unchecked")
-  protected AergoClient supplyAergoClient() {
+  @Test
+  public void testExecuteWithFee() throws Exception {
+    // given
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final TxHash expected = TxHash.of(BytesValue.of(randomUUID().toString().getBytes()));
+    when(mockTxRequester
+        .request(any(AergoClient.class), any(Signer.class), any(TxRequestFunction.class)))
+        .thenReturn(expected);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
+
+    // then
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+        new Class<?>[]{ContractTest.class}, invocationHandler);
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+    final TxHash actual = proxy.executeWithFee(Fee.of(300000L));
+    assertEquals(expected, actual);
+  }
+
+  @Test
+  public void testQuery() throws IOException {
+    // given
+    final int expected = randomUUID().toString().hashCode();
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final ContractResult contractResult = mock(ContractResult.class);
+    when(contractResult.bind(any(Class.class))).thenReturn(expected);
+    when(mockContractOperation.query(any(ContractInvocation.class))).thenReturn(contractResult);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
+
+    // then
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+        new Class<?>[]{ContractTest.class}, invocationHandler);
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+    final int actual = proxy.query();
+    assertEquals(expected, actual);
+  }
+
+  @Test
+  public void shouldThrowErrorIfReturnOnExecution() throws IOException {
+    // given
+    final int expected = randomUUID().toString().hashCode();
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
+
+    // when
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+        new Class<?>[]{ContractTest.class}, invocationHandler);
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+
     try {
-      final ContractOperation mockContractOperation = mock(ContractOperation.class);
-
-      final ContractInterface contractInterface = supplyContractInterface();
-      when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
-          .thenReturn(contractInterface);
-      final ContractTxHash contractTxHash = ContractTxHash.of(BytesValue.EMPTY);
-      when(mockContractOperation
-          .executeTx(any(Signer.class), any(ContractInvocation.class), anyLong(), any(Fee.class)))
-          .thenReturn(contractTxHash);
-      final ContractResult contractResult = mock(ContractResult.class);
-      when(contractResult.bind(any(Class.class))).thenReturn(1);
-      when(mockContractOperation.query(any(ContractInvocation.class))).thenReturn(contractResult);
-
-      final AergoClient mockClient = mock(AergoClient.class);
-      when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
-
-      return mockClient;
+      proxy.invalidExecute();
+      fail();
     } catch (Exception e) {
-      throw new IllegalStateException(e);
+      // then
     }
   }
 
+  @Test
+  public void shouldThrowErrorIfVoidReturnOnQuery() throws IOException {
+    // given
+    final int expected = randomUUID().toString().hashCode();
+    final ContractOperation mockContractOperation = mock(ContractOperation.class);
+    final ContractInterface contractInterface = supplyContractInterface();
+    when(mockContractOperation.getContractInterface(any(ContractAddress.class)))
+        .thenReturn(contractInterface);
+    final AergoClient mockClient = mock(AergoClient.class);
+    when(mockClient.getContractOperation()).thenReturn(mockContractOperation);
+    final TxRequester mockTxRequester = mock(TxRequester.class);
+    final ContractInvocationHandler invocationHandler = new ContractInvocationHandler(
+        contractAddress, mockTxRequester);
+
+    // when
+    final ContractTest proxy = (ContractTest) Proxy.newProxyInstance(classLoader,
+        new Class<?>[]{ContractTest.class}, invocationHandler);
+    final Signer signer = new AergoKeyGenerator().create();
+    invocationHandler.prepareClient(mockClient);
+    invocationHandler.prepareSigner(signer);
+
+    try {
+      proxy.invalidQuery();
+      fail();
+    } catch (Exception e) {
+      // then
+    }
+  }
+
+  @SuppressWarnings("ArraysAsListWithZeroOrOneArgument")
   protected ContractInterface supplyContractInterface() {
     final List<ContractFunction> contractFunctions = asList(
-        new ContractFunction("execute", asList("arg")),
-        new ContractFunction("query", false, true, false)
+        new ContractFunction("executeVoid", Collections.<String>emptyList(), false, false, false),
+        new ContractFunction("executeTxHash", Collections.<String>emptyList(), false, false, false),
+        new ContractFunction("executeWithFee", asList("arg"), false, false, false),
+        new ContractFunction("query", Collections.<String>emptyList(), false, true, false),
+        new ContractFunction("invalidExecute", Collections.<String>emptyList(), false, false,
+            false),
+        new ContractFunction("invalidQuery", Collections.<String>emptyList(), false, true, false)
     );
-    final List<StateVariable> stateVariables = new ArrayList<>();
     return ContractInterface.newBuilder()
         .address(contractAddress)
         .version("1.0")
         .language("lua")
         .functions(contractFunctions)
-        .stateVariables(stateVariables)
+        .stateVariables(Collections.<StateVariable>emptyList())
         .build();
+  }
+
+  private interface ContractTest {
+
+    void executeVoid();
+
+    TxHash executeTxHash(String a, String b);
+
+    TxHash executeWithFee(Fee fee);
+
+    int query();
+
+    Object invalidExecute();
+
+    void invalidQuery();
   }
 
 }