+++ client/wallet/src/main/java/hera/wallet/WalletBuilder.java
@@ -7,6 +7,7 @@ package hera.wallet;
 import hera.annotation.ApiAudience;
 import hera.annotation.ApiStability;
 import hera.api.model.Time;
+import hera.api.model.TryCountAndInterval;
 import hera.client.AergoClient;
 import hera.client.AergoClientBuilder;
 import hera.client.ClientConfiguer;