+++ client/wallet/src/test/java/hera/wallet/WalletApiImplTest.java
@@ -11,7 +11,6 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static org.mockito.Mockito.mock;
 
 import hera.AbstractTestCase;
 import hera.api.model.Aer;
@@ -22,6 +21,7 @@ import hera.api.model.Hash;
 import hera.api.model.RawTransaction;
 import hera.api.model.Signature;
 import hera.api.model.Time;
+import hera.api.model.TryCountAndInterval;
 import hera.client.AergoClient;
 import hera.client.AergoClientBuilder;
 import hera.exception.HerajException;