+++ core/common/src/main/java/hera/api/model/ContractInvocation.java
@@ -31,6 +31,7 @@ public class ContractInvocation implements Payload {
   @Getter
   @NonNull
   @Default
+  @ToString.Exclude
   protected final ContractFunction function = ContractFunction.EMPTY;
 
   @Getter