+++ core/transport/src/main/java/hera/client/TxRequester.java
@@ -2,13 +2,16 @@
  * @copyright defined in LICENSE.txt
  */
 
-package hera.wallet;
+package hera.client;
 
+import hera.annotation.ApiAudience;
+import hera.annotation.ApiStability;
 import hera.api.model.TxHash;
-import hera.client.AergoClient;
 import hera.key.Signer;
 
-interface TxRequester {
+@ApiAudience.Private
+@ApiStability.Unstable
+public interface TxRequester {
 
   TxHash request(AergoClient aergoClient, Signer signer, TxRequestFunction requestFunction)
       throws Exception;